+++ lib/claper_web/live/event_live/event_form_component.ex
@@ -96,8 +96,7 @@ defmodule ClaperWeb.EventLive.EventFormComponent do
 
     static_path =
       Path.join([
-        :code.priv_dir(:claper),
-        "static",
+        get_presentation_storage_dir(),
         "uploads",
         "#{hash}"
       ])
@@ -114,7 +113,7 @@ defmodule ClaperWeb.EventLive.EventFormComponent do
                 "original.#{ext}"
               ])
 
-            # The `static/uploads` directory must exist for `File.cp!/2` to work.
+            # The storage directory must exist for `File.cp!/2` to work.
             File.mkdir_p!(static_path)
 
             File.cp!(path, dest)
@@ -220,6 +219,10 @@ defmodule ClaperWeb.EventLive.EventFormComponent do
 
   defp get_max_file_size() do
     Application.get_env(:claper, :max_file_size)
+  end
+
+  defp get_presentation_storage_dir do
+    Application.get_env(:claper, :presentations) |> Keyword.get(:storage_dir)
   end
 
   def error_to_string(:too_large), do: gettext("Your file is too large")